@@ -1,21 +1,11 @@
 {
   "name": "voyagecanada-frontend",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.1",
-    "@mui/utils": "^5.4.4",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-geocode": "^0.2.3",
-    "react-icons": "^4.3.1",
-    "react-places-autocomplete": "^7.3.0",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
